+++ airbyte-ci/connectors/pipelines/pipelines/airbyte_ci/test/commands.py
@@ -18,7 +18,7 @@ from pipelines.models.contexts.click_pipeline_context import ClickPipelineContex
 @pass_pipeline_context
 @click_ignore_unused_kwargs
 async def test(pipeline_context: ClickPipelineContext):
-    """Runs the tests for the given airbyte-ci package.
+    """Runs the tests for the given airbyte-ci package
 
     Args:
         pipeline_context (ClickPipelineContext): The context object.
@@ -30,7 +30,19 @@ async def test(pipeline_context: ClickPipelineContext):
     logger.info(f"Running tests for {poetry_package_path}")
 
     # The following directories are always mounted because a lot of tests rely on them
-    directories_to_always_mount = [".git", ".github", "docs", "airbyte-integrations", "airbyte-ci", "airbyte-cdk", "pyproject.toml"]
+    directories_to_always_mount = [
+        ".git",  # This is needed as some package tests rely on being in a git repo
+        ".github",
+        "docs",
+        "airbyte-integrations",
+        "airbyte-ci",
+        "airbyte-cdk",
+        "pyproject.toml",
+        "LICENSE_SHORT",
+        "poetry.lock",
+        "spotless-maven-pom.xml",
+        "tools/gradle/codestyle/java-google-style.xml",
+    ]
     directories_to_mount = list(set([poetry_package_path, *directories_to_always_mount]))
 
     pipeline_name = f"Unit tests for {poetry_package_path}"
@@ -63,6 +75,7 @@ async def test(pipeline_context: ClickPipelineContext):
         .with_workdir(f"/airbyte/{poetry_package_path}")
         .with_exec(["poetry", "install"])
         .with_unix_socket("/var/run/docker.sock", dagger_client.host().unix_socket("/var/run/docker.sock"))
+        .with_env_variable("CI", str(pipeline_context.params["is_ci"]))
         .with_exec(["poetry", "run", "pytest", test_directory])
     )