+++ airbyte-integrations/connectors/source-instagram/unit_tests/conftest.py
@@ -35,6 +35,7 @@ def config_fixture():
 def some_config_fixture(account_id):
     return {"start_date": "2021-01-23T00:00:00Z", "access_token": "unknown_token"}
 
+
 @fixture(scope="session", name="some_config_future_date")
 def some_config_future_date_fixture(account_id):
     return {"start_date": "2030-01-23T00:00:00Z", "access_token": "unknown_token"}