+++ airbyte-integrations/connectors/source-salesforce/unit_tests/api_test.py
@@ -8,6 +8,7 @@ import io
 import logging
 import re
 from datetime import datetime
+from typing import List
 from unittest.mock import Mock
 
 import freezegun
@@ -15,6 +16,8 @@ import pendulum
 import pytest
 import requests_mock
 from airbyte_cdk.models import AirbyteStream, ConfiguredAirbyteCatalog, ConfiguredAirbyteStream, DestinationSyncMode, SyncMode, Type
+from airbyte_cdk.sources.streams import Stream
+from airbyte_cdk.sources.streams.concurrent.adapters import StreamFacade
 from airbyte_cdk.utils import AirbyteTracedException
 from conftest import encoding_symbols_parameters, generate_stream
 from requests.exceptions import HTTPError
@@ -25,9 +28,14 @@ from source_salesforce.streams import (
     CSV_FIELD_SIZE_LIMIT,
     BulkIncrementalSalesforceStream,
     BulkSalesforceStream,
+    Describe,
     IncrementalRestSalesforceStream,
     RestSalesforceStream,
+    SalesforceStream,
 )
+
+_ANY_CATALOG = ConfiguredAirbyteCatalog.parse_obj({"streams": []})
+_ANY_CONFIG = {}
 
 
 @pytest.mark.parametrize(
@@ -56,7 +64,7 @@ from source_salesforce.streams import (
 def test_login_authentication_error_handler(
     stream_config, requests_mock, login_status_code, login_json_resp, expected_error_msg, is_config_error
 ):
-    source = SourceSalesforce()
+    source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
     logger = logging.getLogger("airbyte")
     requests_mock.register_uri(
         "POST", "https://login.salesforce.com/services/oauth2/token", json=login_json_resp, status_code=login_status_code
@@ -335,7 +343,7 @@ def test_encoding_symbols(stream_config, stream_api, chunk_size, content_type_he
 def test_check_connection_rate_limit(
     stream_config, login_status_code, login_json_resp, discovery_status_code, discovery_resp_json, expected_error_msg
 ):
-    source = SourceSalesforce()
+    source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
     logger = logging.getLogger("airbyte")
 
     with requests_mock.Mocker() as m:
@@ -372,7 +380,7 @@ def test_rate_limit_bulk(stream_config, stream_api, bulk_catalog, state):
     stream_1.page_size = 6
     stream_1.state_checkpoint_interval = 5
 
-    source = SourceSalesforce()
+    source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
     source.streams = Mock()
     source.streams.return_value = streams
     logger = logging.getLogger("airbyte")
@@ -428,7 +436,7 @@ def test_rate_limit_rest(stream_config, stream_api, rest_catalog, state):
     stream_1.state_checkpoint_interval = 3
     configure_request_params_mock(stream_1, stream_2)
 
-    source = SourceSalesforce()
+    source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
     source.streams = Mock()
     source.streams.return_value = [stream_1, stream_2]
 
@@ -613,7 +621,7 @@ def test_forwarding_sobject_options(stream_config, stream_names, catalog_stream_
                 ],
             },
         )
-        source = SourceSalesforce()
+        source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
         source.catalog = catalog
         streams = source.streams(config=stream_config)
     expected_names = catalog_stream_names if catalog else stream_names
@@ -621,8 +629,101 @@ def test_forwarding_sobject_options(stream_config, stream_names, catalog_stream_
 
     for stream in streams:
         if stream.name != "Describe":
-            assert stream.sobject_options == {"flag1": True, "queryable": True}
+            if isinstance(stream, StreamFacade):
+                assert stream._legacy_stream.sobject_options == {"flag1": True, "queryable": True}
+            else:
+                assert stream.sobject_options == {"flag1": True, "queryable": True}
     return
+
+
+@pytest.mark.parametrize(
+    "stream_names,catalog_stream_names,",
+    (
+        (
+            ["stream_1", "stream_2", "Describe"],
+            None,
+        ),
+        (
+            ["stream_1", "stream_2"],
+            ["stream_1", "stream_2", "Describe"],
+        ),
+        (
+            ["stream_1", "stream_2", "stream_3", "Describe"],
+            ["stream_1", "Describe"],
+        ),
+    ),
+)
+def test_unspecified_and_incremental_streams_are_not_concurrent(stream_config, stream_names, catalog_stream_names) -> None:
+    for stream in _get_streams(stream_config, stream_names, catalog_stream_names, SyncMode.incremental):
+        assert isinstance(stream, (SalesforceStream, Describe))
+
+
+@pytest.mark.parametrize(
+    "stream_names,catalog_stream_names,",
+    (
+        (
+            ["stream_1", "stream_2"],
+            ["stream_1", "stream_2", "Describe"],
+        ),
+        (
+            ["stream_1", "stream_2", "stream_3", "Describe"],
+            ["stream_1", "Describe"],
+        ),
+    ),
+)
+def test_full_refresh_streams_are_concurrent(stream_config, stream_names, catalog_stream_names) -> None:
+    for stream in _get_streams(stream_config, stream_names, catalog_stream_names, SyncMode.full_refresh):
+        assert isinstance(stream, StreamFacade)
+
+
+def _get_streams(stream_config, stream_names, catalog_stream_names, sync_type) -> List[Stream]:
+    sobjects_matcher = re.compile("/sobjects$")
+    token_matcher = re.compile("/token$")
+    describe_matcher = re.compile("/describe$")
+    catalog = None
+    if catalog_stream_names:
+        catalog = ConfiguredAirbyteCatalog(
+            streams=[
+                ConfiguredAirbyteStream(
+                    stream=AirbyteStream(name=catalog_stream_name, supported_sync_modes=[sync_type], json_schema={"type": "object"}),
+                    sync_mode=sync_type,
+                    destination_sync_mode=DestinationSyncMode.overwrite,
+                )
+                for catalog_stream_name in catalog_stream_names
+            ]
+        )
+    with requests_mock.Mocker() as m:
+        m.register_uri("POST", token_matcher, json={"instance_url": "https://fake-url.com", "access_token": "fake-token"})
+        m.register_uri(
+            "GET",
+            describe_matcher,
+            json={
+                "fields": [
+                    {
+                        "name": "field",
+                        "type": "string",
+                    }
+                ]
+            },
+        )
+        m.register_uri(
+            "GET",
+            sobjects_matcher,
+            json={
+                "sobjects": [
+                    {
+                        "name": stream_name,
+                        "flag1": True,
+                        "queryable": True,
+                    }
+                    for stream_name in stream_names
+                    if stream_name != "Describe"
+                ],
+            },
+        )
+        source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
+        source.catalog = catalog
+        return source.streams(config=stream_config)
 
 
 def test_csv_field_size_limit():
@@ -806,7 +907,7 @@ def test_bulk_stream_request_params_states(stream_config_date_format, stream_api
     stream_config_date_format.update({"start_date": "2023-01-01"})
     stream: BulkIncrementalSalesforceStream = generate_stream("Account", stream_config_date_format, stream_api)
 
-    source = SourceSalesforce()
+    source = SourceSalesforce(_ANY_CATALOG, _ANY_CONFIG)
     source.streams = Mock()
     source.streams.return_value = [stream]